chore: refactor cloud functions to a modular structure
